backend/pyproject.toml
[tool.poetry]
name = "trading-platform-backend"
version = "0.1.0"
description = "Algorithmic Trading Platform — FastAPI Backend"
authors = ["Trading Platform"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

# ── Web framework ─────────────────────────────────────────────────────────────
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.27.0" }
python-multipart = "^0.0.9"        # Form data parsing

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy = { extras = ["asyncio"], version = "^2.0.0" }
asyncpg = "^0.29.0"                # Async PostgreSQL driver
alembic = "^1.13.0"               # Database migrations
psycopg2-binary = "^2.9.9"        # Sync driver (needed by Alembic)

# ── Validation & settings ─────────────────────────────────────────────────────
pydantic = "^2.6.0"
pydantic-settings = "^2.2.0"

# ── Redis ─────────────────────────────────────────────────────────────────────
redis = { extras = ["hiredis"], version = "^5.0.0" }   # hiredis = faster parser

# ── Logging ───────────────────────────────────────────────────────────────────
loguru = "^0.7.0"

# ── HTTP client ───────────────────────────────────────────────────────────────
httpx = "^0.27.0"

# ── Data & ML (Phase 2) ───────────────────────────────────────────────────────
pandas = "^2.2.0"
numpy = "^1.26.0"
yfinance = "^0.2.37"
scikit-learn = "^1.4.0"    # Preprocessing + metrics
xgboost = "^2.0.0"         # Gradient boosting classifier
joblib = "^1.3.0"          # Model serialization
statsmodels = "^0.14.0"    # Phase 3: cointegration tests (pairs trading)
scipy = "^1.12.0"          # Phase 4: portfolio optimization (efficient frontier)
# Phase 5: paper trading uses local DB prices — no external API dependency
shap = "^0.44.0"           # Phase 6: SHAP feature explainability for XGBoost
vaderSentiment = "^3.3.2"  # Phase 14: Financial news sentiment scoring (no model download)
alpaca-py = "^0.30.0"      # Phase 13: Alpaca REST API for real-time market quotes
python-jose = {extras = ["cryptography"], version = "^3.3.0"}  # Phase 17: JWT tokens
passlib = {extras = ["bcrypt"], version = "^1.7.4"}            # Phase 17: Password hashing

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.27.0"                  # Test client for FastAPI
black = "^24.0.0"
ruff = "^0.3.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ── Tooling config ────────────────────────────────────────────────────────────
[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]  # pycodestyle + pyflakes + isort + pyupgrade

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
